Enhance limited shell security - see commit for details
In a multi-user environment users could connect to the server vis SFTP.
In doing so the user would be able to break out of their home directory
and view/modify server system files. The idea is to share a server for
private seedboxes with your friends... not give them access to your
total server.

NOTE: Although SFTP is disabled for your additional created users, there
is still secure protocol for connecting on port 5757 via FTP/TLS, i.e;
encrypted connections with self-signed certificate.